+++ ext_tables.sql
@@ -6,7 +6,8 @@ CREATE TABLE tx_t3vstage_domain_model_slide (
   title varchar(255) DEFAULT '' NOT NULL,
   abstract varchar(255) DEFAULT '' NOT NULL,
   text text NOT NULL,
-  image int(11) unsigned DEFAULT '0' NOT NULL,
+  thumbnails int(11) unsigned DEFAULT '0' NOT NULL,
+  images int(11) unsigned DEFAULT '0' NOT NULL,
   link varchar(255) DEFAULT '' NOT NULL,
 
   tstamp int(11) unsigned DEFAULT '0' NOT NULL,