+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `thumbnails` property to `Slide` Model
+* Multiple images per `Slide` Model (breaking change)
 * Updated TypoScript
 * Fixed functional tests
 * Updated AppVeyor configuration